Fix code style issues
