+++ hudi-integ-test/src/main/scala/org/apache/hudi/integ/testsuite/dag/nodes/SparkDeleteNode.scala
@@ -18,12 +18,13 @@
 
 package org.apache.hudi.integ.testsuite.dag.nodes
 
+import org.apache.avro.Schema
+import org.apache.avro.generic.GenericRecord
 import org.apache.hudi.client.WriteStatus
 import org.apache.hudi.config.HoodieWriteConfig
 import org.apache.hudi.integ.testsuite.configuration.DeltaConfig.Config
 import org.apache.hudi.integ.testsuite.dag.ExecutionContext
-import org.apache.hudi.integ.testsuite.utils.SparkUtils
-import org.apache.hudi.{AvroConversionUtils, DataSourceWriteOptions}
+import org.apache.hudi.{AvroConversionUtils, DataSourceWriteOptions, HoodieSparkUtils}
 import org.apache.log4j.LogManager
 import org.apache.spark.rdd.RDD
 import org.apache.spark.sql.SaveMode
@@ -33,7 +34,7 @@ import scala.collection.JavaConverters._
 /**
  * Spark datasource based upsert node
  *
- * @param config1
+ * @param dagNodeConfig DAG node configurations.
  */
 class SparkDeleteNode(dagNodeConfig: Config) extends DagNode[RDD[WriteStatus]] {
 
@@ -50,7 +51,7 @@ class SparkDeleteNode(dagNodeConfig: Config) extends DagNode[RDD[WriteStatus]] {
   override def execute(context: ExecutionContext, curItrCount: Int): Unit = {
     // Deletes can't be fetched using getNextBatch() bcoz, getInsert(schema) from payload will return empty for delete
     // records
-    val genRecsRDD = SparkUtils.generateRecordsForDelete(config, context)
+    val genRecsRDD = generateRecordsForDelete(config, context)
     val inputDF = AvroConversionUtils.createDataFrame(genRecsRDD,
       context.getWriterContext.getHoodieTestSuiteWriter.getSchema,
       context.getWriterContext.getSparkSession)
@@ -64,5 +65,25 @@ class SparkDeleteNode(dagNodeConfig: Config) extends DagNode[RDD[WriteStatus]] {
       .option(HoodieWriteConfig.TBL_NAME.key, context.getHoodieTestSuiteWriter.getCfg.targetTableName)
       .mode(SaveMode.Append)
       .save(context.getHoodieTestSuiteWriter.getWriteConfig.getBasePath)
+  }
+
+  /**
+   * Generates records for delete operations in Spark.
+   *
+   * @param config  Node configs.
+   * @param context The context needed for an execution of a node.
+   * @return Records in {@link RDD}.
+   */
+  private def generateRecordsForDelete(config: Config, context: ExecutionContext): RDD[GenericRecord] = {
+    if (!config.isDisableGenerate) {
+      context.getDeltaGenerator().writeRecords(context.getDeltaGenerator().generateDeletes(config)).count()
+    }
+
+    context.getWriterContext.getHoodieTestSuiteWriter.getNextBatchForDeletes()
+    val pathToRead = context.getWriterContext.getCfg.inputBasePath + "/" + context.getWriterContext.getHoodieTestSuiteWriter.getLastCheckpoint.orElse("")
+
+    val avroDf = context.getWriterContext.getSparkSession.read.format("avro").load(pathToRead)
+    HoodieSparkUtils.createRdd(avroDf, "testStructName", "testNamespace", false,
+      org.apache.hudi.common.util.Option.of(new Schema.Parser().parse(context.getWriterContext.getHoodieTestSuiteWriter.getSchema)))
   }
 }
